% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{addSheet}
\alias{addSheet}
\title{Add a new sheet}
\usage{
addSheet(proxy, name, active = TRUE)
}
\arguments{
\item{proxy}{The x-spreadsheet proxy object}

\item{name}{The name of the new sheet}

\item{active}{Whether the new sheet should be active,
defaults to TRUE}
}
\description{
Adds a new sheet to the spreadsheet,
requires a x-spreadsheet proxy to exist.
}
