% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{cellText}
\alias{cellText}
\title{Set the value in a cell}
\usage{
cellText(proxy, row, col, text, sheetIndex, triggerChange = TRUE)
}
\arguments{
\item{proxy}{The x-spreadsheet proxy object}

\item{row}{The row index of the cell}

\item{col}{The column index of the cell}

\item{text}{The text to set the cell to}

\item{sheetIndex}{The index of the sheet to set the cell in,
uses R 1-based indexing.}

\item{triggerChange}{Whether to trigger a change event,
defaults to TRUE, will trigger an update to the _change
input value}
}
\description{
Sets the text of a cell, requires
a x-spreadsheet proxy to interact with and valid arguments.
Indices are expected to be 1-based (R).
}
