% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-spreadsheet.R
\name{cellTypeConversion}
\alias{cellTypeConversion}
\title{Helper function to convert a cell to the correct type}
\usage{
cellTypeConversion(x)
}
\arguments{
\item{x}{A list with the following structure:
\itemize{
\item text: the text to convert
\item type: the type to convert to
originating from a JSON object returned by x-spreadsheet
}}
}
\description{
Helper function to convert a cell to the correct type
}
