% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameToList.R
\name{dataFrameToList}
\alias{dataFrameToList}
\title{dataFrameToList}
\usage{
dataFrameToList(inputData)
}
\arguments{
\item{inputData}{A list of dataframes where each list entry has a name. This list is converted into the right format for x-spreadsheet. Each list name is converted into a sheetname together with its relevant data.}
}
\description{
dataFrameToList
}
