% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-spreadsheet.R
\name{dfToxSpreadsheet}
\alias{dfToxSpreadsheet}
\title{Convert a data.frame to x-spreadsheet format}
\usage{
dfToxSpreadsheet(df, headerRow = TRUE)
}
\description{
Convert a data.frame to x-spreadsheet format,
while also adding data types in JSON format for later
conversion back to R.
}
