% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{getData}
\alias{getData}
\title{Get the data object}
\usage{
getData(proxy)
}
\arguments{
\item{proxy}{The x-spreadsheet proxy object}
}
\description{
Retrieves the data object by
sending a call to the client, which then
sends the data to
input[\link{paste0(proxy, "_data")}]

By retrieving the data object from the client,
you can get the current data in the spreadsheet, including
all formatting and other custom settings.

TO DO: add logic to:
load the spreadsheet via the getData data object
add logic to parse the data object into a list containing data.frames
}
