% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{invokeRemote}
\alias{invokeRemote}
\title{Allow to send calls to the client from a R Shiny session}
\usage{
invokeRemote(
  proxy,
  method = c("addSheet", "reRender", "deleteSheet", "loadData", "getData"),
  args = list()
)
}
\arguments{
\item{proxy}{the x-spreadsheet proxy object}

\item{method}{the method to call,
should be a valid API public method from
https://hondrytravis.com/x-spreadsheet-doc/
for which the R -> client communication is sensible
(some API methods are not included as
these would send data from the client to R)}
}
\description{
Allow to send calls to the client from a R Shiny session
}
