% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{reRender}
\alias{reRender}
\title{Rerenders the spreadsheet}
\usage{
reRender(proxy)
}
\arguments{
\item{proxy}{The x-spreadsheet proxy object}
}
\description{
This function is most likely not needed
as the rerendering logic is handled already by the
internal spreadsheet functions.
}
