% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{runExamples}
\alias{runExamples}
\title{Run xSpreadsheet Example Applications}
\usage{
runExamples(example = NA, port = 3000)
}
\arguments{
\item{example}{The name of the example to run, or \code{NA} (the default) to
list the available examples.}
}
\description{
Launch xSpreadsheet example applications,
and optionally, your system's web browser.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  # List all available examples
  runExamples()

  # Run one of the examples
  runExamples("01_spreadSheet")

  # Print the directory containing the code for all examples
  system.file("examples", package="xSpreadsheet")
}
}
