% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RXSpreadsheet.R
\name{RXSpreadsheet}
\alias{RXSpreadsheet}
\title{Load the x-spreadsheet javascript spreadsheet editor into Shiny}
\usage{
RXSpreadsheet(data, options = NULL, elementId = NULL)
}
\arguments{
\item{data}{A list object that follows the x-spreadsheet format. This list object can be generated in two ways: use a list generated by input$<example>_RXSpreadsheetData or, to load existing data into the spreadsheet, convert a list of dataframes with DataFrameToList.}

\item{options}{A list object that allows you to interact with the x-spreadsheet API directly, for all options check https://github.com/myliang/x-spreadsheet}

\item{elementId}{htmlwidget elementid}
}
\description{
This function creates an HTML widget to allow R Shiny to work with the
x-spreadsheet library (https://github.com/myliang/x-spreadsheet).
}
\details{
The widget heavily relies on conversion between lists and json.

To maintain the state of the spreadsheet editor one can store input$<example>_RXSpreadsheetData and load it again, see runExample().
}
