% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-spreadsheet.R
\name{spreadSheet}
\alias{spreadSheet}
\title{x-spreadsheet htmlwidget}
\usage{
spreadSheet(data, options = NULL, elementId = NULL)
}
\arguments{
\item{data}{The data to render inside the spreadsheet, the following
options are valid:
\itemize{
\item NULL, this will result in an empty spreadsheet with a single sheet
named "Sheet1"
\item A single data.frame object, this will result in a single sheet
spreadsheet where the name of the first sheet is "Sheet1"
\item A named list where each list entry contains a data.frame object,
a named sheet is generated in correspondence with the named list
}}

\item{options}{A list containing table options/settings, set at
table initialisation.
The list should match the JSON structure below, e.g.: list(
"mode" = "edit", showToolbar = TRUE, ...), JSON:
{
mode: 'edit', // edit | read
showToolbar: true,
showGrid: true,
showContextmenu: true,
view: {
height: () => document.documentElement.clientHeight,
width: () => document.documentElement.clientWidth,
},
row: {
len: 100,
height: 25,
},
col: {
len: 26,
width: 100,
indexWidth: 60,
minWidth: 60,
},
style: {
bgcolor: '#ffffff',
align: 'left',
valign: 'middle',
textwrap: false,
strike: false,
underline: false,
color: '#0a0a0a',
font: {
name: 'Helvetica',
size: 10,
bold: false,
italic: false,
},
},
}}

\item{elementId}{htmlwidget elementid}
}
\description{
Initialises the x-spreadsheet htmlwidget, allows to
create one or more sheets and pass options.
}
