% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-spreadsheet.R
\name{spreadsheetListToDf}
\alias{spreadsheetListToDf}
\title{Convert x-spreadsheet data to a list of data.frames}
\usage{
spreadsheetListToDf(data, headerRow = TRUE)
}
\arguments{
\item{data}{The data to convert, a list with the following structure
\itemize{
\item name
\item rows, which contains "cells", which contains entries with "text"
other fields are ignored for now
(customise this function and jsso it carries the following metadata:
\item field type for conversion
\item column names Y/N
)
originating from a JSON object returned by x-spreadsheet,
containing one entry per sheet, with the following structure:
{
name: "Sheet1",
rows: {
0: {
cells: {
{text: "A1"},
}
}
}
}
}}

\item{headerRow}{Whether the first row contains column names,
TO DO: automatically detect this by carrying extra metadata with
each sheet}
}
\description{
Convert x-spreadsheet data to a list of data.frames,
this function turns the data structure generated by the JS library
(and parsed with jsonlite::fromJSON(simpifyVector = FALSE)))
into a list of data.frames, one per sheet. It is the inverse of
listToSpreadsheet.
}
