% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{storeState}
\alias{storeState}
\title{Store the data object}
\usage{
storeState(data, path)
}
\arguments{
\item{data}{The data object to store,
must be a stringified JSON generated via
getData and received as input$[]}

\item{path}{The path to store the data object at}
}
\value{
TRUE if successfull
}
\description{
Stores the data object by
parsing the data (a stringified JSON)
and storing it as a JSON object
}
