% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-spreadsheet.R
\name{xSpreadsheetListToDf}
\alias{xSpreadsheetListToDf}
\title{Convert x-spreadsheet data to a list of data.frames}
\usage{
xSpreadsheetListToDf(data, headerRow = TRUE)
}
\arguments{
\item{headerRow}{Whether the first row contains column names,
TO DO: automatically detect this by carrying extra metadata with
each sheet}
}
\description{
Convert x-spreadsheet data to a list of data.frames,
this function turns the data structure generated by the JS library
(and parsed with jsonlite::fromJSON(simpifyVector = FALSE)))
into a list of data.frames, one per sheet. It is the inverse of
listToxSpreadsheet.
}
