% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{xSpreadsheetProxy}
\alias{xSpreadsheetProxy}
\title{Create a x-spreadsheet proxy object}
\usage{
xSpreadsheetProxy(
  outputId,
  session = shiny::getDefaultReactiveDomain(),
  deferUntilFlush = TRUE
)
}
\arguments{
\item{outputId}{the output variable to read the x-spreadsheet}

\item{session}{the shiny session}

\item{deferUntilFlush}{whether to defer the creation of the proxy
until the next flush cycle}
}
\description{
Create a x-spreadsheet proxy object
}
