% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{deleteSheet}
\alias{deleteSheet}
\title{Delete the current active sheet}
\usage{
deleteSheet(proxy)
}
\arguments{
\item{proxy}{The x-spreadsheet proxy object}
}
\description{
Seems broken in x-spreadsheet,
not exporting until fixed.
}
