% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{deleteSheet}
\alias{deleteSheet}
\title{Delete a sheet}
\usage{
deleteSheet(proxy, sheetIndex = NULL)
}
\arguments{
\item{proxy}{The x-spreadsheet proxy object}

\item{sheetIndex}{The index of the sheet to delete,
uses R 1-based indexing. If left NULL, the active sheet
is deleted.}
}
\description{
Deletes a sheet from the spreadsheet,
defaults to deleting the active sheet. Only
deletes sheets if at least two sheets are present.
}
