% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-spreadsheet.R
\name{processInputData}
\alias{processInputData}
\title{Process data for RXSpreadsheet}
\usage{
processInputData(data)
}
\description{
Process data for RXSpreadsheet,
this function turns the different structures allowed for data into the same
format that is expected by the client side logic.
}
